+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/nodes/input/InputIndexOfNode.java
@@ -44,26 +44,18 @@ import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.nodes.Node;
-import com.oracle.truffle.api.strings.JavaStringUtils;
 import com.oracle.truffle.api.strings.TruffleString;
 import com.oracle.truffle.regex.tregex.string.Encodings.Encoding;
 
 public abstract class InputIndexOfNode extends Node {
 
-    public abstract int execute(Object input, int fromIndex, int maxIndex, int[] ranges, Encoding encoding);
+    public abstract int execute(TruffleString input, int fromIndex, int maxIndex, TruffleString.ByteIndexOfCodePointSetNode.CodePointSet codePointSet, Encoding encoding);
 
     @Specialization
-    public int doString(String input, int fromIndex, int maxIndex, int[] ranges, @SuppressWarnings("unused") Encoding encoding,
-                    @Cached JavaStringUtils.CharIndexOfCodePointSetNode indexOfNode) {
-        CompilerAsserts.partialEvaluationConstant(ranges);
-        return indexOfNode.execute(input, fromIndex, maxIndex, ranges);
-    }
-
-    @Specialization
-    public int doTString(TruffleString input, int fromIndex, int maxIndex, int[] ranges, Encoding encoding,
+    public int doTString(TruffleString input, int fromIndex, int maxIndex, TruffleString.ByteIndexOfCodePointSetNode.CodePointSet codePointSet, Encoding encoding,
                     @Cached TruffleString.ByteIndexOfCodePointSetNode indexOfNode) {
-        CompilerAsserts.partialEvaluationConstant(ranges);
-        return indexOfNode.execute(input, fromIndex << encoding.getStride(), maxIndex << encoding.getStride(), ranges, encoding.getTStringEncoding()) >> encoding.getStride();
+        CompilerAsserts.partialEvaluationConstant(codePointSet);
+        return indexOfNode.execute(input, fromIndex << encoding.getStride(), maxIndex << encoding.getStride(), codePointSet) >> encoding.getStride();
     }
 
     public static InputIndexOfNode create() {