+++ regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/nodes/nfa/TRegexLiteralLookAroundExecutorNode.java
@@ -119,13 +119,13 @@ public final class TRegexLiteralLookAroundExecutorNode extends TRegexBacktracker
 
     @TruffleBoundary
     @Override
-    public TRegexExecutorLocals createLocals(Object input, int fromIndex, int index, int maxIndex) {
+    public TRegexExecutorLocals createLocals(TruffleString input, int fromIndex, int index, int maxIndex) {
         throw new UnsupportedOperationException();
     }
 
     @ExplodeLoop
     @Override
-    public Object execute(VirtualFrame frame, TRegexExecutorLocals abstractLocals, TruffleString.CodeRange codeRange, boolean tString) {
+    public Object execute(VirtualFrame frame, TRegexExecutorLocals abstractLocals, TruffleString.CodeRange codeRange) {
         TRegexBacktrackingNFAExecutorLocals locals = (TRegexBacktrackingNFAExecutorLocals) abstractLocals;
         for (int i = 0; i < matchers.length; i++) {
             if (!inputHasNext(locals) || !matchers[i].match(inputReadAndDecode(locals, codeRange))) {